login works sometimes, but doesnt update with useEffect when expected
